pythonexample/requirements.txt
cassandra-driver==3.24.0
lz4==3.1.3 